+++ examples/rust/get_started/src/lib.rs
@@ -4,13 +4,11 @@ pub use echoer::*;
 mod hop;
 pub use hop::*;
 
-mod attribute_access_control;
 mod credentials;
 mod logger;
 mod project;
 mod token;
 
-pub use attribute_access_control::*;
 pub use credentials::*;
 pub use logger::*;
 pub use project::*;
